remove user styles and consolidate around settings styles for consistency
