--- src/app/components/user/user.module.css
@@ -1,151 +1,0 @@
-.user-settings-wrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  overflow: auto;
-
-  .user-content {
-    display: flex;
-    flex: 1 1 auto;
-    flex-direction: column;
-    overflow: auto;
-    padding: 16px 16px 32px;
-
-    .appstore-badges {
-      display: flex;
-      gap: 16px;
-
-      a {
-        /* stylelint-disable-next-line scale-unlimited/declaration-strict-value */
-        color: unset;
-        display: inline-flex;
-        overflow: hidden;
-
-        svg {
-          &.appstore-apple {
-            height: 40px;
-            width: 120px;
-          }
-
-          &.appstore-play {
-            height: 40px;
-            width: 135px;
-          }
-        }
-      }
-    }
-
-    .user-setting-details-wrapper {
-      max-width: var(--columnWidthLarge);
-      width: 100%;
-
-      .user-setting-content-container {
-        background-color: var(--color-white-100);
-        border-radius: var(--border-radius-default);
-        padding: 16px;
-
-        .user-setting-content-container-title {
-          align-items: center;
-          display: flex;
-          justify-content: space-between;
-          margin: 0 0 8px;
-          min-height: 34px;
-
-          > span {
-            @mixin typography-subtitle1;
-            align-items: center;
-            display: flex;
-            flex: 1 1 auto;
-            gap: 2px;
-          }
-        }
-
-        .user-setting-content-container-inner {
-          background-color: var(--color-beige-93);
-          border-radius: var(--border-radius-medium);
-          padding: 16px;
-
-          + .user-setting-content-container-inner {
-            margin: 16px 0 0;
-          }
-        }
-
-        .user-setting-content-container-inner-accent {
-          background: var(--color-beige-86);
-          border-radius: var(--border-radius-small);
-          padding: 16px;
-
-          + .user-setting-content-container-inner-accent {
-            margin: 16px 0 0;
-          }
-        }
-
-        &.user-info-primary {
-          flex: 3;
-
-          .user-info-name {
-            align-items: center;
-            display: flex;
-            gap: 8px;
-            margin-bottom: 8px;
-          }
-
-          .contact-info {
-            align-items: center;
-            color: var(--color-gray-37);
-            display: flex;
-            flex-flow: wrap row;
-            gap: 8px;
-            margin: 0 0 16px;
-          }
-        }
-
-        + .user-setting-content-container {
-          margin: 16px 0 0;
-        }
-      }
-
-      .user-setting-content-list {
-        li {
-          align-items: flex-start;
-          border-bottom: solid 1px var(--color-gray-88);
-          display: flex;
-          gap: 16px;
-          margin: 0 0 8px;
-          padding: 8px 0 16px;
-
-          &.setting-content-list-language {
-            align-items: center;
-          }
-
-          &:last-of-type {
-            border-bottom: 0;
-            margin: 0;
-            padding: 8px 0 0;
-          }
-        }
-      }
-
-      .user-info-edit {
-        display: flex;
-        gap: 32px;
-
-        .user-info-primary {
-          width: 100%;
-        }
-      }
-
-      .StyledAvatarEditButton {
-        margin: 8px 0 0;
-        width: 100%;
-      }
-
-      .user-info-avatar {
-        flex: 0;
-        flex-shrink: 0;
-        justify-content: center;
-        margin: 0 16px 0 0;
-      }
-    }
-  }
-}
